+++ docker/Dockerfile
@@ -5,7 +5,7 @@ WORKDIR /src
 
 RUN go install google.golang.org/protobuf/cmd/protoc-gen-go@latest
 RUN go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@latest
-RUN go install github.com/bufbuild/buf/cmd/buf@latest
+RUN go install github.com/bufbuild/buf/cmd/buf@v1.13.1
 
 COPY daemon/proto .
 COPY buf.gen.yaml .
@@ -41,6 +41,7 @@ ARG TARGETARCH
 COPY daemon/ daemon/
 COPY api/ api/
 COPY plugin/ plugin/
+COPY utils/ utils/
 COPY --from=proto_base /src/ .
